# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Agency Owner Operating System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /payments:
    get:
      operationId: listPayments
      tags: [payments]
      summary: List payments
      parameters:
        - name: clientId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of payments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"
    post:
      operationId: createPayment
      tags: [payments]
      summary: Create a payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /payments/{id}:
    get:
      operationId: getPayment
      tags: [payments]
      summary: Get a payment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
        "404":
          description: Not found
    patch:
      operationId: updatePayment
      tags: [payments]
      summary: Update a payment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentUpdate"
      responses:
        "200":
          description: Updated payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
    delete:
      operationId: deletePayment
      tags: [payments]
      summary: Delete a payment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Documents ───────────────────────────────────────────────────────────
...
  /notes:
    get:
      operationId: listNotes
      tags: [notes]
      summary: List notes
      parameters:
        - name: clientId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of notes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Note"
    post:
      operationId: createNote
      tags: [notes]
      summary: Create a note
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
  /notes/{id}:
    get:
      operationId: getNote
      tags: [notes]
      summary: Get a note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
        "404":
          description: Not found
    patch:
      operationId: updateNote
      tags: [notes]
      summary: Update a note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteUpdate"
      responses:
        "200":
          description: Updated note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
    delete:
      operationId: deleteNote
      tags: [notes]
      summary: Delete a note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Meetings ────────────────────────────────────────────────────────────
...
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List tasks
      parameters:
        - name: clientId
          in: query
          schema:
            type: integer
        - name: projectId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Timeline ────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Payment:
      type: object
      required: [id, clientId, invoiceNumber, amount, status, createdAt]
      properties:
        id:
          type: integer
        clientId:
          type: integer
        clientName:
          type: ["string", "null"]
        projectId:
          type: ["integer", "null"]
        invoiceNumber:
          type: string
        amount:
          type: number
        status:
          type: string
          enum: [paid, pending, overdue, cancelled]
        dueDate:
          type: ["string", "null"]
        paidDate:
          type: ["string", "null"]
        paymentMethod:
          type: ["string", "null"]
        remainingBalance:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    PaymentInput:
      type: object
      required: [clientId, invoiceNumber, amount, status]
      properties:
        clientId:
          type: integer
        projectId:
          type: integer
        invoiceNumber:
          type: string
        amount:
          type: number
        status:
          type: string
          enum: [paid, pending, overdue, cancelled]
        dueDate:
          type: string
        paidDate:
          type: string
        paymentMethod:
          type: string
        remainingBalance:
          type: number
        notes:
          type: string
    PaymentUpdate:
      type: object
      properties:
        invoiceNumber:
          type: string
        amount:
          type: number
        status:
          type: string
          enum: [paid, pending, overdue, cancelled]
        dueDate:
          type: string
        paidDate:
          type: string
        paymentMethod:
          type: string
        remainingBalance:
          type: number
        notes:
          type: string

    # ─── Document ────────────────────────────────────────────────────────────
...
    Note:
      type: object
      required: [id, content, createdAt]
      properties:
        id:
          type: integer
        clientId:
          type: ["integer", "null"]
        projectId:
          type: ["integer", "null"]
        clientName:
          type: ["string", "null"]
        content:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
    NoteInput:
      type: object
      required: [content]
      properties:
        clientId:
          type: integer
        projectId:
          type: integer
        content:
          type: string
    NoteUpdate:
      type: object
      properties:
        content:
          type: string

    # ─── Meeting ─────────────────────────────────────────────────────────────
...
    Task:
      type: object
      required: [id, title, priority, status, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        priority:
          type: string
          enum: [low, medium, high, urgent]
        status:
          type: string
          enum: [todo, in_progress, done]
        deadline:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        clientId:
          type: ["integer", "null"]
        projectId:
          type: ["integer", "null"]
        clientName:
          type: ["string", "null"]
        projectName:
          type: ["string", "null"]
        createdAt:
          type: string
    TaskInput:
      type: object
      required: [title, priority, status]
      properties:
        title:
          type: string
        priority:
          type: string
          enum: [low, medium, high, urgent]
        status:
          type: string
          enum: [todo, in_progress, done]
        deadline:
          type: string
        notes:
          type: string
        clientId:
          type: integer
        projectId:
          type: integer
    TaskUpdate:
      type: object
      properties:
        title:
          type: string
        priority:
          type: string
          enum: [low, medium, high, urgent]
        status:
          type: string
          enum: [todo, in_progress, done]
        deadline:
          type: string
        notes:
          type: string
        clientId:
          type: integer
        projectId:
          type: integer

    # ─── Timeline ────────────────────────────────────────────────────────────
...